requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
click==8.1.3
comm==0.1.2
debugpy==1.6.4
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
importlib-metadata==5.2.0
ipykernel==6.19.4
ipython==8.7.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==7.4.8
jupyter_core==5.1.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
packaging==22.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.0
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pyzmq==24.0.1
six==1.16.0
SQLAlchemy==1.4.45
stack-data==0.6.2
tornado==6.2
traitlets==5.8.0
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.11.0
